+++ tests/testthat/test-python-initialize.R
@@ -13,3 +13,10 @@ test_that("sys.executable points to the correct python", {
 
 })
 
+
+test_that("__main__ initialzed with only 'r'", {
+
+  expect_identical(
+    callr::r(function() names(reticulate::import_main())),
+    "r")
+})